+++ gauss_bot/matrices/ops_matrices.py
@@ -2,370 +2,12 @@ from fractions import Fraction
 from typing import Tuple, List, Dict, overload
 from copy import deepcopy
 
-from utils import limpiar_pantalla
+from gauss_bot.matrices.matriz import Matriz
+from gauss_bot.matrices.sistema_ecuaciones import SistemaEcuaciones
+from gauss_bot.utils import limpiar_pantalla
 
 # type annotations:
 DictMatrices = Dict[str, "Matriz"]
-Validacion = Tuple[bool, int]
-
-
-class Matriz:
-    def __init__(self, aumentada: bool, filas: int, columnas: int, valores: List[List[Fraction]] = []):
-        self.aumentada = aumentada
-        self.filas = filas
-        self.columnas = columnas
-
-        if valores == []:
-            self.valores = [[Fraction(0) for _ in range(columnas)] for _ in range(filas)]
-        else:
-            self.valores = valores
-
-    @overload
-    def __getitem__(self, indice_fila: int) -> List[Fraction]: ...
-
-    @overload
-    def __getitem__(self, indices: Tuple[int, int]) -> Fraction: ...
-
-    def __getitem__(self, indice):
-        if isinstance(indice, int):
-            return self.valores[indice]
-        elif isinstance(indice, tuple) and len(indice) == 2:
-            fila, columna = indice
-            return self.valores[fila][columna]
-        else:
-            raise TypeError("Índice inválido")
-
-    def __setitem__(self, indices: Tuple[int, int], valor: Fraction) -> None:
-        fila, columna = indices
-        self.valores[fila][columna] = valor
-
-    def es_matriz_cero(self) -> bool:
-        return all(all(x == Fraction(0) for x in fila) for fila in self.valores)
-
-    def get_mat_str(self) -> str:
-        max_len = max(
-            len(str(self.valores[i][j].limit_denominator(100)))
-            for i in range(self.filas)
-            for j in range(self.columnas)
-        )
-
-        matriz = ""
-        for i in range(self.filas):
-            for j in range(self.columnas):
-                var = str(self.valores[i][j].limit_denominator(100)).center(max_len)
-                if j == 0 and j == self.columnas - 1:
-                    matriz += f"( {var} )"
-                elif j == 0:
-                    matriz += f"( {var}, "
-                elif j == self.columnas - 2 and self.aumentada:
-                    matriz += f"{var} | "
-                elif j == self.columnas - 1:
-                    matriz += f"{var} )"
-                else:
-                    matriz += f"{var}, "
-            matriz += "\n"
-        return matriz
-
-    def agregar_fila(self, fila: List[Fraction]) -> None:
-        if len(fila) != self.columnas:
-            raise ValueError("La longitud de la fila debe coincidir con el número de columnas")
-        self.valores.append(fila)
-        self.filas += 1
-
-    def agregar_columna(self, columna: List[Fraction]) -> None:
-        if len(columna) != self.filas:
-            raise ValueError("La longitud de la columna debe coincidir con el número de filas")
-        for i in range(self.filas):
-            self.valores[i].append(columna[i])
-        self.columnas += 1
-
-    def eliminar_fila(self, indice_fila: int) -> None:
-        if 0 <= indice_fila < self.filas:
-            self.valores.pop(indice_fila)
-            self.filas -= 1
-        else:
-            raise IndexError("Índice de fila fuera de rango")
-
-    def eliminar_columna(self, indice_columna: int) -> None:
-        if 0 <= indice_columna < self.columnas:
-            for fila in self.valores:
-                fila.pop(indice_columna)
-            self.columnas -= 1
-        else:
-            raise IndexError("Índice de columna fuera de rango")
-
-    def validar_consistencia(self) -> Validacion:
-        if self.valores == [] or self.es_matriz_cero():
-            return (True, -1)
-        
-        for i in range(self.filas):
-            # validar forma de 0 = b (donde b != 0)
-            if [0 for _ in range(len(self.valores[0]) - 1)] == self.valores[i][:-1] and self.valores[i][-1] != 0:
-                return (False, i)
-
-        return (True, -1)
-
-    def validar_escalonada(self) -> bool:
-        fila_cero = [0 for _ in range(self.columnas)]
-        entrada_anterior = -1
-
-        if self.valores == [] or self.es_matriz_cero():
-            return False
-
-        # buscar filas no-cero despues de una fila cero, significa que no es escalonada
-        for i in range(self.filas):
-            if self.valores[i] != fila_cero:
-                continue
-            if any(self.valores[j] != fila_cero for j in range(i + 1, self.filas)):
-                return False
-
-        # validar entradas principales
-        for i in range(self.filas):
-            try:
-                entrada_actual = next(j for j in range(self.columnas - 1) if self.valores[i][j] != 0)
-            except StopIteration:
-                continue
-            if entrada_actual <= entrada_anterior:
-                return False
-            if any(self.valores[k][entrada_actual] != 0 for k in range(i + 1, self.filas)):
-                return False
-            entrada_anterior = entrada_actual
-
-        return True
-
-    def validar_escalonada_reducida(self) -> bool:
-        if not self.validar_escalonada():  # si no es escalonada, no puede ser escalonada reducida
-            return False
-
-        # validar entradas principales
-        for i in range(self.filas):
-            try:
-                entrada_principal = next(j for j in range(self.columnas - 1) if self.valores[i][j] == 1)
-            except StopIteration:
-                continue
-            if any(self.valores[k][entrada_principal] != 0 and k != i for k in range(self.filas)):
-                return False
-
-        return True
-
-    def sumar_mats(self, mat2: "Matriz") -> "Matriz":
-        if self.filas != mat2.filas or self.columnas != mat2.columnas:
-            raise ValueError("Las matrices deben tener las mismas dimensiones")
-        M_sumada = [[self.valores[i][j] + mat2.valores[i][j] for j in range(self.columnas)] for i in range(self.filas)]
-        return Matriz(self.aumentada, self.filas, self.columnas, M_sumada)
-
-    def restar_mats(self, mat2: "Matriz") -> "Matriz":
-        if self.filas != mat2.filas or self.columnas != mat2.columnas:
-            raise ValueError("Las matrices deben tener las mismas dimensiones")
-        M_resta = [[self.valores[i][j] - mat2.valores[i][j] for j in range(self.columnas)] for i in range(self.filas)]
-        return Matriz(self.aumentada, self.filas, self.columnas, M_resta)
-
-    def mult_escalar(self, escalar: Fraction) -> "Matriz":
-        M_multiplicada = [[escalar * valor for valor in fila] for fila in self.valores]
-        return Matriz(self.aumentada, self.filas, self.columnas, M_multiplicada)
-
-    def mult_matrices(self, mat2: "Matriz") -> "Matriz":
-        if self.columnas != mat2.filas:
-            raise ValueError("El número de columnas de la primera matriz debe ser igual al número de filas de la segunda matriz")
-
-        M_mult = [[Fraction(0) for _ in range(mat2.columnas)] for _ in range(self.filas)]
-        for i in range(self.filas):
-            for j in range(mat2.columnas):
-                for k in range(self.columnas):
-                    M_mult[i][j] += self.valores[i][k] * mat2.valores[k][j]
-        
-        return Matriz(self.aumentada, self.filas, mat2.columnas, M_mult)
-
-    def transponer(self) -> "Matriz":
-        M_t = [[self.valores[j][i] for j in range(self.filas)] for i in range(self.columnas)]
-        return Matriz(self.aumentada, self.columnas, self.filas, M_t)
-
-
-class SistemaEcuaciones:
-    def __init__(self, matriz: Matriz) -> None:
-        self.matriz = matriz
-        self.filas = matriz.filas
-        self.columnas = matriz.columnas
-
-        self.respuesta: List[str] = []
-        self.procedimiento: List[str] = []
-
-    def resolver_sistema(self) -> None:
-        self.reducir_matriz()
-        if self.matriz.es_matriz_cero():
-            return None
-
-        libres = self.encontrar_variables_libres()
-        if self.matriz.validar_escalonada_reducida() and not libres:  # solucion unica:
-            self.imprimir_soluciones(unica=True, libres=[], validacion=(True, -1))
-            return None
-
-        # solucion general:
-        self.imprimir_soluciones(unica=False, libres=libres, validacion=(True, -1))
-        return None
-
-    def reducir_matriz(self) -> None:
-        test_inicial = self.matriz.validar_consistencia()
-        if not test_inicial[0]:
-            self.imprimir_soluciones(unica=False, libres=[], validacion=test_inicial)
-            return None
-        elif self.matriz.validar_escalonada_reducida():
-            self.procedimiento.append("\nMatriz ya esta en su forma escalonada reducida!\n")
-            return None
-
-        self.procedimiento.append("\nMatriz inicial:\n")
-        self.procedimiento.append(self.matriz.get_mat_str())
-        self.procedimiento.append("\nReduciendo la matriz a su forma escalonada:\n")
-
-        fila_actual = 0
-        for j in range(self.columnas - 1):  # encontrar pivotes y eliminar elementos debajo
-            fila_pivote = None
-            maximo = Fraction(0)
-
-            # buscar la entrada con el valor mas grande para usarla como pivote
-            for i in range(fila_actual, self.filas):
-                if abs(self.matriz[i, j]) > maximo:
-                    maximo = abs(self.matriz[i, j])
-                    fila_pivote = i
-
-            # si las variables siguen iguales, no hay pivote y se pasa a la siguiente columna
-            if fila_pivote is None or maximo == 0:
-                continue
-
-            # siempre trabajar con la fila pivote
-            if fila_pivote != fila_actual:
-                self.matriz.valores[fila_actual], self.matriz.valores[fila_pivote] = self.matriz.valores[fila_pivote], self.matriz.valores[fila_actual]
-                self.procedimiento.append(f"\nF{fila_actual+1} <==> F{fila_pivote+1}\n")
-                self.procedimiento.append(self.matriz.get_mat_str())
-
-            pivote = self.matriz[fila_actual][j]
-            if pivote != 1 and pivote != 0:  # hacer el pivote igual a 1
-                for k in range(self.columnas):
-                    self.matriz[fila_actual][k] /= pivote
-                if pivote == -1:
-                    self.procedimiento.append(f"\nF{fila_actual+1} => -F{fila_actual+1}\n")
-                else:
-                    self.procedimiento.append(f"\nF{fila_actual+1} => F{fila_actual+1} / {str(pivote.limit_denominator(100))}\n")
-
-                self.procedimiento.append(self.matriz.get_mat_str())
-
-            # eliminar los elementos debajo del pivote
-            for f in range(fila_actual + 1, self.filas):
-                factor = self.matriz[f][j]
-                if factor == 0:
-                    continue
-                for k in range(self.columnas):
-                    self.matriz[f][k] -= factor * self.matriz[fila_actual][k]
-
-                self.procedimiento.append(f"\nF{fila_actual+1} => F{fila_actual+1} - ({str(factor.limit_denominator(100))} * F{fila_pivote+1})\n")
-                self.procedimiento.append(self.matriz.get_mat_str())
-
-            fila_actual += 1
-
-        for i in reversed(range(self.filas)):  # eliminar elementos arriba de pivotes
-            try:
-                columna_pivote = self.matriz[i].index(next(filter(lambda x: x != 0, self.matriz[i])))
-            except StopIteration:
-                columna_pivote = None
-            if columna_pivote is None:  # no hay entrada principal
-                continue
-
-            # eliminar los elementos encima del pivote
-            for f in reversed(range(i)):
-                factor = self.matriz[f][columna_pivote]
-                if factor == 0:
-                    continue
-                for k in range(self.columnas):
-                    self.matriz[f][k] -= factor * self.matriz[i][k]
-
-                self.procedimiento.append(f"\nF{f+1} => F{f+1} - ({str(factor.limit_denominator(100))} * F{i+1})\n")
-                self.procedimiento.append(self.matriz.get_mat_str())
-
-        test_final = self.matriz.validar_consistencia()
-        if not test_final[0]:
-            self.imprimir_soluciones(unica=False, libres=[], validacion=test_final)
-            return None
-        return None
-
-    def encontrar_variables_libres(self) -> List[int]:
-        entradas = []
-
-        # encontrar entradas principales
-        for i in range(self.matriz.filas):
-            try:
-                entradas.append(next(j for j in range(self.matriz.columnas - 1) if self.matriz[i][j] != 0))
-            except StopIteration:
-                continue
-
-        # si no son entradas principales, son variables libres
-        return [x for x in range(self.matriz.columnas - 1) if x not in entradas]
-
-    def despejar_variables(self, libres: List[int]) -> List[str]:
-        ecuaciones = []
-        for x in libres:
-            ecuaciones.append(f"| X{x+1} es libre\n")
-
-        for i in range(self.filas):
-            for j in range(self.columnas - 1):
-                if self.matriz[i, j] == 0:
-                    continue
-
-                constante = self.matriz[i, -1].limit_denominator(100)
-                expresion = str(constante) if constante != 0 else ""
-                for k in range(j + 1, self.columnas - 1):
-                    if self.matriz[i, k] == 0:
-                        continue
-
-                    coeficiente = -self.matriz[i, k].limit_denominator(100)
-                    signo = "+" if coeficiente >= 0 else "-"
-                    if expresion:
-                        signo = f" {signo} "
-                    else:
-                        signo = "" if signo == "+" else "-"
-
-                    variable = (
-                        f"{abs(coeficiente)}*X{k+1}"
-                        if abs(coeficiente) != 1
-                        else f"X{k+1}"
-                    )
-
-                    expresion += f"{signo}{variable}"
-
-                if expresion == "":
-                    expresion = "0"
-
-                ecuaciones.append(f"| X{j+1} = {expresion}\n")
-                break
-
-        ecuaciones.sort(key=lambda x: x[3])
-        return ecuaciones
-
-    def imprimir_soluciones(self, unica: bool, libres: List[int], validacion: Validacion) -> None:
-        solucion, fila_inconsistente = validacion
-        if not solucion and fila_inconsistente != -1:
-            self.respuesta.append(f"\n| En F{fila_inconsistente+1}: 0 != {str(self.matriz[fila_inconsistente, -1].limit_denominator(100))}\n")
-            self.respuesta.append("| Sistema es inconsistente!\n")
-            return None
-
-        elif unica:
-            solucion_trivial = all(self.matriz[i, -1] == 0 for i in range(self.filas))
-            mensaje_solucion = "trivial" if solucion_trivial else "no trivial"
-
-            self.respuesta.append(f"\n| Solución {mensaje_solucion} encontrada:\n")
-            for i in range(self.filas):
-                if all(x == 0 for x in self.matriz[i]):
-                    continue
-                self.respuesta.append(f"| X{i+1} = {str(self.matriz[i, -1].limit_denominator(100))}\n")
-            return None
-
-        ecuaciones = self.despejar_variables(libres)
-        self.respuesta.append("\n| Sistema no tiene solución única!\n")
-        self.respuesta.append("| Solución general encontrada:\n")
-        for linea in ecuaciones:
-            self.respuesta.append(linea)
-
-        return None
-
 
 class OperacionesMatrices:
     def __init__(self, default_mats: DictMatrices = {}) -> None:
@@ -747,3 +389,4 @@ class OperacionesMatrices:
         if input_usuario not in opciones:
             return -1
         return 1 if input_usuario == "s" else 0
+
